# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sentinel AML API
  version: 1.1.0
  description: Static API contract for the implementation in this repository. Basic
    authentication is stateless. Backend-configured users are admin, analyst and viewer;
    credentials come from environment variables and are never included here. All list
    responses mask customer identifiers and sensitive narrative fields. Numeric customer
    IDs support the authorized by-id detail endpoint. VIEWER detail views remain redacted.
    CSV and JSON batch submission returns 202 with a durable job to poll. Detection
    occurs atomically for each processed transaction. No external
    SAR filing API is implemented.
servers:
- url: http://localhost:8080/api/v1
  description: Local development; use HTTPS for deployment
security:
- basicAuth: []
paths:
...
  "/customers":
    get:
      operationId: listCustomers
      summary: List customers
      tags:
      - Customers
      description: 'Allowed roles: ADMIN, ANALYST, VIEWER.'
      x-roles:
      - ADMIN
      - ANALYST
      - VIEWER
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageCustomerResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: Repeat for multiple sorts, e.g. sort=riskScore,desc.
    post:
      operationId: createCustomer
      summary: Create or ingest customer
      tags:
      - Customers
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CustomerResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CustomerRequest"
  "/customers/{customerId}":
    get:
      operationId: getCustomer
      summary: Get customer detail
      tags:
      - Customers
      description: 'Allowed roles: ADMIN, ANALYST.'
      x-roles:
      - ADMIN
      - ANALYST
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CustomerResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: string
...
  "/customers/by-id/{id}":
    get:
      operationId: getCustomerByInternalId
      summary: Get authorized customer detail by numeric internal ID
      tags:
      - Customers
      description: 'Allowed roles: ADMIN, ANALYST.'
      x-roles:
      - ADMIN
      - ANALYST
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CustomerResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
  "/customers/{customerId}/accounts":
    get:
      operationId: listCustomerAccounts
      summary: List a customer’s accounts
      tags:
      - Customers
      description: 'Allowed roles: ADMIN, ANALYST, VIEWER.'
      x-roles:
      - ADMIN
      - ANALYST
      - VIEWER
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PageAccountResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: customerId
        in: path
        required: true
        schema:
          type: string
      - name: page
        in: query
        schema:
          type: integer
          minimum: 0
          default: 0
      - name: size
        in: query
        schema:
          type: integer
          minimum: 1
          default: 20
      - name: sort
        in: query
        schema:
          type: array
          items:
            type: string
        style: form
        explode: true
        description: Repeat for multiple sorts, e.g. sort=riskScore,desc.
...
  "/rules":
    get:
      operationId: listRules
      summary: List configured detection rules
      tags:
      - Rules
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/RuleSettings"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
  "/rules/{code}":
    put:
      operationId: updateRule
      summary: Update an enabled rule and configuration
      tags:
      - Rules
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/RuleSettings"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RuleUpdate"
  "/settings/exchange-rates":
    get:
      operationId: listExchangeRate
      summary: List exchange rates
      tags:
      - Settings
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ExchangeRate"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
    put:
      operationId: upsertExchangeRate
      summary: Create or update exchange rates
      tags:
      - Settings
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExchangeRate"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RateUpdate"
  "/settings/high-risk-jurisdictions":
    get:
      operationId: listHighRiskJurisdiction
      summary: List high risk jurisdictions
      tags:
      - Settings
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/HighRiskJurisdiction"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
    put:
      operationId: upsertHighRiskJurisdiction
      summary: Create or update high risk jurisdictions
      tags:
      - Settings
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HighRiskJurisdiction"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/JurisdictionUpdate"
  "/settings/sanctioned-counterparties":
    get:
      operationId: listSanctionedCounterparty
      summary: List sanctioned counterparties
      tags:
      - Settings
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/SanctionedCounterparty"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
    put:
      operationId: upsertSanctionedCounterparty
      summary: Create or update sanctioned counterparties
      tags:
      - Settings
      description: 'Allowed roles: ADMIN.'
      x-roles:
      - ADMIN
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SanctionedCounterparty"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        default:
          "$ref": "#/components/responses/Problem"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CounterpartyUpdate"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    KycStatus:
      type: string
      enum:
      - PENDING
      - VERIFIED
      - REJECTED
      - EXPIRED
...
    AccountStatus:
      type: string
      enum:
      - ACTIVE
      - DORMANT
      - CLOSED
      - FROZEN
      - SUSPENDED
    RiskRating:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
...
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountId:
          type: string
        customerId:
          type: string
        accountType:
          type: string
        riskRating:
          "$ref": "#/components/schemas/RiskRating"
        accountStatus:
          "$ref": "#/components/schemas/AccountStatus"
        currency:
          type: string
        openDate:
          type: string
          format: date
        closeDate:
          type: string
          format: date
        branchCode:
          type: string
        branchCity:
          type: string
        currentBalance:
          type: number
        avgMonthlyBalance6m:
          type: number
        creditLimit:
          type: number
        creditUtilizationPct:
          type: number
        overdraftEnabled:
          type: boolean
        cardType:
          type: string
        jointAccount:
          type: boolean
        numLinkedDevices:
          type: integer
          format: int32
        mobileBankingEnrolled:
          type: boolean
        lastLoginDate:
          type: string
          format: date
        avgMonthlyTxnCount:
          type: integer
          format: int32
        accountTier:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CustomerRequest:
      type: object
      properties:
        customerId:
          type: string
          maxLength: 64
          minLength: 1
        firstName:
          type: string
          maxLength: 100
          minLength: 1
        lastName:
          type: string
          maxLength: 100
          minLength: 1
        gender:
          type: string
          maxLength: 20
        dateOfBirth:
          type: string
          format: date
        email:
          type: string
          maxLength: 255
          format: email
        phoneNumber:
          type: string
          maxLength: 50
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 100
        country:
          type: string
          maxLength: 3
        postalCode:
          type: string
          maxLength: 20
        occupation:
          type: string
          maxLength: 150
        annualIncome:
          type: number
        maritalStatus:
          type: string
          maxLength: 30
        educationLevel:
          type: string
          maxLength: 50
        employmentStatus:
          type: string
          maxLength: 30
        customerSince:
          type: string
          format: date
        customerSegment:
          type: string
          maxLength: 30
        kycStatus:
          "$ref": "#/components/schemas/KycStatus"
        riskRating:
          "$ref": "#/components/schemas/RiskRating"
        politicallyExposed:
          type: boolean
        preferredChannel:
          type: string
          maxLength: 50
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        numComplaintsLastYear:
          type: integer
          format: int32
      required:
      - customerId
      - firstName
      - lastName
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        gender:
          type: string
        dateOfBirth:
          type: string
          format: date
        email:
          type: string
        phoneNumber:
          type: string
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        occupation:
          type: string
        annualIncome:
          type: number
        maritalStatus:
          type: string
        educationLevel:
          type: string
        employmentStatus:
          type: string
        customerSince:
          type: string
          format: date
        customerSegment:
          type: string
        kycStatus:
          "$ref": "#/components/schemas/KycStatus"
        riskRating:
          "$ref": "#/components/schemas/RiskRating"
        politicallyExposed:
          type: boolean
        preferredChannel:
          type: string
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        numComplaintsLastYear:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      description: List views mask customerId and personal fields. ADMIN/ANALYST can
        retrieve authorized detail by numeric id or external customerId.
    PageCustomerResponse:
      type: object
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      - last
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/CustomerResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
    PageAccountResponse:
      type: object
      required:
      - content
      - page
      - size
      - totalElements
      - totalPages
      - last
      properties:
        content:
          type: array
          items:
            "$ref": "#/components/schemas/AccountResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        last:
          type: boolean
...
    RuleSettings:
      type: object
      properties:
        ruleCode:
          type: string
          enum:
          - CTR
          - STRUCTURING
          - RAPID_MOVEMENT
          - HIGH_RISK_JURISDICTION
          - BEHAVIORAL_DEVIATION
          - ROUND_NUMBER
        enabled:
          type: boolean
        configuration:
          type: object
          additionalProperties: true
      required:
      - ruleCode
      - enabled
      - configuration
    RuleUpdate:
      type: object
      properties:
        enabled:
          type: boolean
        configuration:
          type: object
          additionalProperties: true
      required:
      - enabled
      - configuration
      description: Only known settings are accepted. CTR and HIGH_RISK_JURISDICTION
        cannot be disabled. Monetary rule thresholds specify currency (USD by default);
        server normalizes with effective FX rates.
    RateUpdate:
      type: object
      properties:
        fromCurrency:
          type: string
          pattern: "^[A-Za-z]{3}$"
        toCurrency:
          type: string
          pattern: "^[A-Za-z]{3}$"
        rate:
          type: number
          minimum: 1.0e-08
        effectiveFrom:
          type: string
          format: date-time
      required:
      - fromCurrency
      - toCurrency
      - rate
      - effectiveFrom
    JurisdictionUpdate:
      type: object
      properties:
        countryCode:
          type: string
          pattern: "^[A-Za-z]{2,3}$"
        description:
          type: string
          maxLength: 255
        sanctioned:
          type: boolean
        effectiveFrom:
          type: string
          format: date-time
        effectiveTo:
          type: string
          format: date-time
          nullable: true
      required:
      - countryCode
      - effectiveFrom
    CounterpartyUpdate:
      type: object
      properties:
        identifier:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 500
        enabled:
          type: boolean
      required:
      - identifier
    ExchangeRate:
      allOf:
      - "$ref": "#/components/schemas/RateUpdate"
      - type: object
        properties:
          id:
            type: integer
            format: int64
    HighRiskJurisdiction:
      allOf:
      - "$ref": "#/components/schemas/JurisdictionUpdate"
      - type: object
        properties:
          id:
            type: integer
            format: int64
    SanctionedCounterparty:
      "$ref": "#/components/schemas/CounterpartyUpdate"
...
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        errorType:
          type: string
        sourceReference:
          type: string
  responses:
    Unauthorized:
      description: Missing or invalid Basic credentials
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated role cannot access this resource
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Problem:
      description: Request failure; RFC 9457 problem details, including 400 validation,
        404 missing resource, 409 duplicate/conflict, 422 missing parent, or 500 unexpected
        failure.
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"